refactor: move parsing helpers in their own module
